Renamed CheckboxRoot to Checkbox
